fix: moved banner to separate components
