+++ app/src/components/mode-specific/desktop/InterceptTraffic/WebTraffic/TrafficTableV2/index.jsx
@@ -580,6 +580,12 @@ const CurrentTrafficTable = ({
     [dispatch]
   );
 
+  useEffect(() => {
+    if (disableFilters) {
+      setIsFiltersCollapsed(true);
+    }
+  }, [disableFilters]);
+
   // IMP: Keep this in the end to wait for other useEffects to run first
   useEffect(() => {
     mounted.current = true;
